two model finish .
I create 18 endPoint with CRUD

+++ src/main/java/com/example/program/Model/InsurancePackage.java
@@ -19,7 +19,7 @@ public class InsurancePackage {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     @NotNull(message = "Should be not null")
-    private Integer insurancePrice;
+    private Double insurancePrice;
     private String duration = "Year";
     @NotEmpty(message = "Should be not empty")
     @Column(columnDefinition = "varchar(15) NOT NULL unique check(name = 'Third party insurance' or name = 'full insurance')")
